more tests added and dtos new added, new price style

+++ src/main/java/com/example/demo/repository/entity/ProductEntity.java
@@ -1,7 +1,6 @@
 package com.example.demo.repository.entity;
 
-import io.swagger.v3.oas.annotations.Hidden;
-import io.swagger.v3.oas.annotations.media.Schema;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,16 +9,17 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Document( collection = "products" )
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-
 public class ProductEntity {
 
     @Id
     private String productID;
     private String name;
-    private Double price;
+    private BigDecimal price;
 }
